refactor(deployments): refactoring of appenvs to apps
