+++ src/app/space/create/deployments/deployments.component.html
@@ -1,4 +1,4 @@
 <div id="apps">
-  <deployments-appenvs [environments]="environments" [applications]="applications"></deployments-appenvs>
+  <deployments-apps [environments]="environments" [applications]="applications"></deployments-apps>
   <deployments-resource-usage [environments]="environments"></deployments-resource-usage>
 </div>
